Modify random_integer to take an actual probability distribution
The probability distribution that random_integer wants given is not
intuitive.  What we want to give to it is {0.25, 0.25, 0.5}.  This means that
there is a 25% chance it will return 1, a 25% chance it will return 2, and a
50% chance it will return 3.  Instead, what it wanted was {0.25, 0.5, 1},
which is a lot less intuitive.

Originally, it generated a random number, and looped through prob_distrib.
It looped through prob_distrib until the random number was greater than the
value at the current index of prob_distrib.  Now, after checking the value,
it subtracts the current value of prob_distrib from the random number,
allowing the programmer to input an actual probability distribution.