Add list_delete method
The list_delete method deletes the (logically) first record from list "list"
with a value "value" for attribute "attribute."  It places the attributes of
the deleted record in the transfer array.

For error condition (e.g., there is no matching record in the list), it
returns a value of 0.  Otherwise, it returns a value of 1.