Add more sim logic. Definitely not complete yet.
Forgot to add the different loading times to the different plane types.
Now it runs, but the taxi is misbehaving.  When finishing deberthing a plane,
it checks to see if the runway is empty, or if the berths are full, and if
so, it just goes back to the berths and is idle again.  But the problem is,
there will be a couple hundred planes in the runway queue and it will still
decide to just mosey on back to the berths and wait.  The jerk.

Definitely need to fix.