Add input file, and add global struct G to handle input values.
Program so far reads input file and loads values into G.  Format of input
file is simply <string> <string> <value>.  It doesn't matter what the strings
are, just the order of the values.

Program does not use G for values yet, it still uses the #define constants in
the header.